backend/pyproject.toml
[tool.poetry]
name = "win-loss-counter"
version = "0.1.0"
description = "Socket backend for win-loss-counter"
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
flask = "^3.0.3"
flask-socketio = "^5.3.6"
python-decouple = "^3.8"
flask-cors = "^4.0.1"
uwsgi = {version="^2.0.26", platform="linux"}
supervisor = {version="^4.2.5", platform="linux"}
gevent = {version="^24.2.1", platform="linux"}

[tool.poetry.group.dev.dependencies]
ni-python-styleguide = "^0.4.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
